Update linear actuator numbers
The linear actuators did not operate in the same way as intended, so I swapped the addresses of 2 and 3. Additionally, I removed the 'wait for serial begin' in the setup, because that made it so that the robot would only work if the serial monitor was opened up (i.e. it wouldn't work with just the controller being used to operate it).